requirements.txt
hfutils>=0.2.2
hbutils>=0.9.0
huggingface_hub>=0.22
tqdm
requests
click>=7
pillow